gws_demo/pyproject.toml
[project]
name = "gws-demo"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.191.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
]
